Improves Docker development and production setup
Configures separate .env files for development and production environments, and ensures they are copied during initialization.

This allows for environment-specific configurations for development and production.

Also updates the docker-compose setup to create separate containers for development and production modes and adds `down-dev` and `down-prod` Makefile targets to remove them properly.